Leopard build fix after r96530.
* rendering/svg/RenderSVGResourcePattern.cpp:
(WebCore::RenderSVGResourcePattern::applyResource):



Canonical link: https://commits.webkit.org/85313@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@96551 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/rendering/svg/RenderSVGResourcePattern.cpp
@@ -123,7 +123,8 @@ bool RenderSVGResourcePattern::applyResource(RenderObject* object, RenderStyle* 
         FloatRect clampedAbsoluteTileBoundaries;
 
         // Scale the tile size to match the scale level of the patternTransform.
-        absoluteTileBoundaries.scale(m_attributes.patternTransform().xScale(), m_attributes.patternTransform().yScale());
+        absoluteTileBoundaries.scale(static_cast<float>(m_attributes.patternTransform().xScale()),
+            static_cast<float>(m_attributes.patternTransform().yScale()));
 
         // Build tile image.
         OwnPtr<ImageBuffer> tileImage = createTileImage(m_attributes, tileBoundaries, absoluteTileBoundaries, tileImageTransform, clampedAbsoluteTileBoundaries);
